requirements.txt
requests==2.27.1
python-dotenv==0.19.2
chevron==0.14.0
tldextract==3.1.2